# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GameInCome",
        "description": "Description of API accessor methods for GameInCome",
        "contact": {
            "name": "Support",
            "email": "karlossan@me.com"
        },
        "license": {
            "name": "License 1.0 (Demo)",
            "url": "L5_SWAGGER_CONST_HOST"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://game-back.test/api/v1",
            "description": "Production API Server"
        },
        {
            "url": "https://game-back.karlossan.com/api/v1",
            "description": "Stage API Server"
        }
    ],
    "paths": {
...
        "/personal-area/appeals/{appealType}": {
            "get": {
                "tags": [
                    "Personal Area: Appeals"
                ],
                "summary": "Personal Area Appeals",
                "operationId": "personal_area_appeals",
                "parameters": [
                    {
                        "name": "appealType",
                        "in": "path",
                        "description": "general / dispute / partnership / tech-support",
                        "required": true,
                        "example": "general"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowAppealsResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Произошла ошибка: Способ оплаты временно недоступен, попробуйте перезагрузить страницу, если ошибка остаётся, то обратитесь в поддержку, Спасибо!"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Нет прав для выполнения запроса"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personal Area: Appeals"
                ],
                "summary": "Create Appeals",
                "operationId": "personal_area_appeals_create",
                "parameters": [
                    {
                        "name": "appealType",
                        "in": "path",
                        "description": "general / dispute / partnership / tech-support",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAppealsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Error: Bad Request"
                    },
                    "401": {
                        "description": "Error: Unauthenticated"
                    },
                    "403": {
                        "description": "Error: Forbidden / Недостаточно прав для выполнения запроса"
                    },
                    "404": {
                        "description": "Error: Resource Not Found"
                    },
                    "409": {
                        "description": "Error: Conflict / Не указаны или указаны некорректно обязательные для запроса dispute параметры: itemType и keyId"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/personal-area/appeals/{appealType}/{appealId}/messages": {
            "get": {
                "tags": [
                    "Personal Area: Appeals"
                ],
                "summary": "Personal Area Appeal Messages",
                "operationId": "personal_area_appeals_messages",
                "parameters": [
                    {
                        "name": "appealType",
                        "in": "path",
                        "description": "general / dispute / partnership / tech-support",
                        "required": true
                    },
                    {
                        "name": "appealId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowAppealMessagesResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Произошла ошибка: Способ оплаты временно недоступен, попробуйте перезагрузить страницу, если ошибка остаётся, то обратитесь в поддержку, Спасибо!"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Нет прав для выполнения запроса"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/authorization/login": {
            "post": {
                "tags": [
                    "OLD_Authentication & Authorization"
                ],
                "summary": "validates an account",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Вы ввели неверные данные, попробуйте снова!"
                    },
                    "400": {
                        "description": "Произошла ошибка во время входа, личный токен не был сгенерирован, попробуйте перезагрузить страницу, если ошибка остаётся, то обратитесь в поддержку, Спасибо!"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateAppealsRequest": {
                "required": [
                    "theme",
                    "text"
                ],
                "properties": {
                    "theme": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "itemType": {
                        "description": "Only for disputes: games / software / swiches / cases",
                        "type": "string",
                        "example": null
                    },
                    "keyId": {
                        "description": "Only for disputes",
                        "type": "string",
                        "example": null
                    },
                    "image": {
                        "description": "Only for disputes not required",
                        "type": "image",
                        "example": null
                    }
                },
                "type": "object"
            },
...
            "ShowAppealMessagesResource": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "userName": {
                                    "type": "string"
                                },
                                "userSurname": {
                                    "type": "string"
                                },
                                "userAvatar": {
                                    "type": "string"
                                },
                                "text": {
                                    "type": "string"
                                },
                                "image": {
                                    "type": "string"
                                },
                                "isAdmin": {
                                    "type": "boolean"
                                },
                                "createdAt": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "integer",
                        "example": "200"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShowAppealMessagesResource"
                }
            },
            "ShowAppealsResource": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "appealId": {
                                    "type": "string"
                                },
                                "appealNumber": {
                                    "type": "number"
                                },
                                "appealTheme": {
                                    "type": "string"
                                },
                                "appealAnswered": {
                                    "type": "string"
                                },
                                "appealClosedAt": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "integer",
                        "example": "200"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShowAppealsResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer 35|a7aabf6c23d1...)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}